+++ scripts/disabled_tests/common/models.py
@@ -1,4 +1,5 @@
-from typing import TypedDict
+import dataclasses as datacls
+from typing import TypedDict, ClassVar
 
 
 class Scheme(TypedDict):
@@ -12,3 +13,18 @@ class Scheme(TypedDict):
 
 class SchemeWithStatus(Scheme):
     ISSUE_TRACKER_STATUS: str
+
+
+@datacls.dataclass
+class JdkInfo:
+    version: str
+    implementation: str
+
+    DEFAULT_VERSION: ClassVar = 'all'
+    DEFAULT_IMPLEMENTATION: ClassVar = 'hotspot'
+
+    def __post_init__(self):
+        # allow default values to be set
+        # by passing `None` in the constructor
+        self.version = self.version or self.DEFAULT_VERSION
+        self.implementation = self.implementation or self.DEFAULT_IMPLEMENTATION
